# contracts/api-surface.openapi.yaml
openapi: 3.1.0
info:
  title: ShanHaiEdu Core Integration Surface
  version: 3.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    当前可调用的核心联调合同。operationId 必须与后端运行时 OpenAPI 双向一致；
    尚未实现的接口只允许进入 planned-api-surface.openapi.yaml。
...
servers:
  - url: /api/v2
security:
  - cookieAuth: []
paths:
...
  /projects/{project_id}/materials:
    get:
      summary: 查询项目教材材料列表
      operationId: listProjectTextbookMaterials
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Newest-first source material page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceMaterialListEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
  /projects/{project_id}/material-scope/artifact:
    get:
      summary: 恢复项目当前教材范围及审批事实
      operationId: getMaterialScopeArtifact
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Exact material-scope aggregate from PostgreSQL
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MaterialScopeArtifactEnvelope'}
        '4XX': {$ref: '#/components/responses/Error'}
  /projects/{project_id}/material-scope/versions:
    post:
      summary: 创建或修订项目教材范围版本
      operationId: createMaterialScopeVersion
      security:
        - BrowserOrigin: []
          CsrfToken: []
          cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaterialScopeVersionRequest'
      responses:
        '201':
          description: Submitted immutable material-scope version created or replayed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
  /projects/{project_id}/materials/uploads:
    post:
      summary: 创建教材私有直传会话
      operationId: createMaterialUploadSession
      security:
        - BrowserOrigin: []
          CsrfToken: []
          cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadSessionRequest'
      responses:
        '201':
          description: Presigned upload session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSessionEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
  /projects/{project_id}/materials/{material_id}/confirm:
    post:
      summary: 确认教材对象并创建后台任务
      operationId: confirmMaterialUpload
      security:
        - BrowserOrigin: []
          CsrfToken: []
          cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/MaterialId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmUploadRequest'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '4XX':
          $ref: '#/components/responses/Error'
...
  /projects/{project_id}/lesson-division/artifact:
    get:
      summary: 恢复项目当前课时划分及审批事实
      operationId: getLessonDivisionArtifact
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Exact lesson-division aggregate from PostgreSQL
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LessonDivisionArtifactEnvelope'}
        '4XX': {$ref: '#/components/responses/Error'}
  /projects/{project_id}/lesson-division/artifact-versions/{artifact_version_id}/quality-validations:
    post:
      summary: 对项目的课时划分版本执行质量检查
      operationId: startLessonDivisionQualityValidation
      security:
        - BrowserOrigin: []
          CsrfToken: []
          cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactVersionId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '202':
          description: Exact lesson-division quality NodeRun accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AcceptedNodeRunEnvelope'}
        '4XX': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    BrowserOrigin:
      type: apiKey
      in: header
      name: Origin
      description: >-
        Browser-controlled request origin. User agents attach this header;
        application callers must not synthesize it from user input.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        Server-issued token bound to the current session and injected by the web client.
    cookieAuth:
      type: apiKey
      in: cookie
      name: shanhai_session
      description: >-
        Server-issued Secure HttpOnly session resolved through the authentication adapter.
        Organization and user identity headers are not accepted.
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 64
        maxLength: 64
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    MaterialId:
      name: material_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    ArtifactVersionId:
      name: artifact_version_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 128}
...
    PageCursor:
      name: page[cursor]
      in: query
      required: false
      schema: {type: string}
    PageLimit:
      name: page[limit]
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: './error-envelope.schema.json'
    AcceptedJob:
      description: Long-running job accepted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AcceptedJobEnvelope'
  schemas:
...
    CreateUploadSessionRequest:
      type: object
      additionalProperties: false
      required: [filename, media_type, size_bytes, sha256]
      properties:
        filename: {type: string, minLength: 1, maxLength: 255}
        media_type: {type: string, minLength: 1, maxLength: 255}
        size_bytes: {type: integer, minimum: 1}
        sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
    UploadSession:
      type: object
      additionalProperties: false
      required: [upload_session_id, material_id, upload_url, method, required_headers, expires_at]
      properties:
        upload_session_id: {type: string, format: uuid}
        material_id: {type: string, format: uuid}
        upload_url: {type: string, format: uri}
        method: {const: PUT}
        required_headers:
          type: object
          additionalProperties: {type: string}
        expires_at: {type: string, format: date-time}
    UploadSessionEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: '#/components/schemas/UploadSession'}
        request_id: {type: string, minLength: 1}
    ConfirmUploadRequest:
      type: object
      additionalProperties: false
      required: [upload_session_id, etag, size_bytes, sha256]
      properties:
        upload_session_id: {type: string, format: uuid}
        etag: {type: string, minLength: 1, maxLength: 255}
        size_bytes: {type: integer, minimum: 1}
        sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
...
    SourceMaterial:
      type: object
      additionalProperties: false
      required:
        - id
        - project_id
        - material_kind
        - file_asset_id
        - original_filename
        - mime_type
        - upload_status
        - confirmed_at
        - created_at
        - updated_at
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        material_kind: {type: string, minLength: 1}
        file_asset_id: {type: [string, 'null'], format: uuid}
        original_filename: {type: string, minLength: 1}
        mime_type: {type: string, minLength: 1}
        upload_status: {enum: [pending_upload, confirmed, rejected]}
        confirmed_at: {type: [string, 'null'], format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    SourceMaterialListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta, request_id]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              items: {$ref: '#/components/schemas/SourceMaterial'}
        meta:
          type: object
          additionalProperties: false
          required: [next_cursor]
          properties:
            next_cursor: {type: [string, 'null'], format: uuid}
        request_id: {type: string, minLength: 1}
...
    AcceptedNodeRunEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [node_run_id, status, events_url]
          properties:
            node_run_id: {type: string, format: uuid}
            status: {$ref: './workflow-node-status.schema.json'}
            events_url: {type: string, minLength: 1}
        request_id: {type: string, minLength: 1}
...
    AcceptedJobEnvelope:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [job_id, status, events_url]
          properties:
            job_id: {type: string, format: uuid}
            status: {enum: [created, queued, running]}
            events_url: {type: string}
        request_id: {type: string}
...
    CreateMaterialScopeVersionRequest:
      type: object
      additionalProperties: false
      required:
        - source_material_id
        - material_parse_version_id
        - page_start
        - page_end
      properties:
        source_material_id: {type: string, format: uuid}
        material_parse_version_id: {type: string, format: uuid}
        page_start: {type: integer, minimum: 1}
        page_end: {type: integer, minimum: 1}
...
    ArtifactDraft:
      type: object
      additionalProperties: false
      required:
        - id
        - draft_branch
        - content
        - validation_report
        - based_on_version_id
        - autosaved_at
        - lock_version
      properties:
        id: {type: string, format: uuid}
        draft_branch: {type: string}
        content: {type: object, additionalProperties: true}
        validation_report: {type: object}
        based_on_version_id: {type: [string, 'null'], format: uuid}
        autosaved_at: {type: string, format: date-time}
        lock_version: {type: integer, minimum: 1}
    ArtifactVersion:
      type: object
      additionalProperties: false
      required:
        - id
        - version_no
        - content
        - content_hash
        - render_summary
        - source_kind
        - source_node_run_id
        - context_snapshot_id
        - prompt_snapshot_id
        - validation_report
        - created_at
        - created_by
      properties:
        id: {type: string, format: uuid}
        version_no: {type: integer, minimum: 1}
        content: {type: object, additionalProperties: true}
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        render_summary: {type: object}
        source_kind: {enum: [manual, model, import, system]}
        source_node_run_id: {type: [string, 'null'], format: uuid}
        context_snapshot_id: {type: [string, 'null'], format: uuid}
        prompt_snapshot_id: {type: [string, 'null'], format: uuid}
        validation_report: {type: object}
        created_at: {type: string, format: date-time}
        created_by: {type: string, format: uuid}
    Artifact:
      type: object
      additionalProperties: false
      required:
        - id
        - project_id
        - lesson_unit_id
        - branch_key
        - artifact_key
        - artifact_type
        - content_definition_version_id
        - status
        - stale_reason
        - lock_version
        - current_draft
        - current_submitted_version
        - current_approved_version
        - created_at
        - updated_at
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        lesson_unit_id: {type: [string, 'null'], format: uuid}
        branch_key: {type: string}
        artifact_key: {type: string}
        artifact_type: {type: string}
        content_definition_version_id: {type: string, format: uuid}
        status: {enum: [draft, in_review, approved, stale, archived]}
        stale_reason:
          oneOf:
            - {$ref: './artifact-stale-reason.schema.json'}
            - {type: 'null'}
        lock_version: {type: integer, minimum: 1}
        current_draft:
          oneOf:
            - {$ref: '#/components/schemas/ArtifactDraft'}
            - {type: 'null'}
        current_submitted_version:
          oneOf:
            - {$ref: '#/components/schemas/ArtifactVersion'}
            - {type: 'null'}
        current_approved_version:
          oneOf:
            - {$ref: '#/components/schemas/ArtifactVersion'}
            - {type: 'null'}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Approval:
      type: object
      additionalProperties: false
      required:
        - id
        - artifact_version_id
        - action
        - actor_type
        - actor_user_id
        - comment
        - quality_evidence
        - policy_snapshot
        - created_at
      properties:
        id: {type: string, format: uuid}
        artifact_version_id: {type: string, format: uuid}
        action: {enum: [submit, approve, request_changes, revoke, accept_stale]}
        actor_type: {enum: [user, system]}
        actor_user_id: {type: [string, 'null'], format: uuid}
        comment: {type: [string, 'null']}
        quality_evidence: {type: object}
        policy_snapshot: {type: object}
        created_at: {type: string, format: date-time}
    ArtifactEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: '#/components/schemas/Artifact'}
        request_id: {type: string, minLength: 1}
    LessonPlanQualityReport:
      type: object
      additionalProperties: false
      required:
        - id
        - artifact_version_id
        - validate_node_run_id
        - conclusion
        - findings
        - evidence_hash
        - created_at
      properties:
        id: {type: string, format: uuid}
        artifact_version_id: {type: string, format: uuid}
        validate_node_run_id: {type: string, format: uuid}
        conclusion: {enum: [passed, failed]}
        findings: {type: array, items: {type: object}}
        evidence_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        created_at: {type: string, format: date-time}
...
    MaterialScopeArtifactEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [artifact, latest_approval]
          properties:
            artifact:
              oneOf:
                - {$ref: '#/components/schemas/Artifact'}
                - {type: 'null'}
            latest_approval:
              oneOf:
                - {$ref: '#/components/schemas/Approval'}
                - {type: 'null'}
        request_id: {type: string, minLength: 1}
    LessonDivisionArtifactEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [artifact, quality_report, latest_approval]
          properties:
            artifact:
              oneOf:
                - {$ref: '#/components/schemas/Artifact'}
                - {type: 'null'}
            quality_report:
              oneOf:
                - {$ref: '#/components/schemas/LessonPlanQualityReport'}
                - {type: 'null'}
            latest_approval:
              oneOf:
                - {$ref: '#/components/schemas/Approval'}
                - {type: 'null'}
        request_id: {type: string, minLength: 1}
...